aider: Adjust `heroku.yml` to run `init-db.sh` before deploying and update `Dockerfile` to start the server with `npm start`.
# Aider chat conversation:

USER: Ok, now we need to make any adjustments respective of the fact that we are leveraging heroku.yml and not docker-compose.yml.

ASSISTANT: To adjust for the use of `heroku.yml` instead of `docker-compose.yml`, we need to make sure that the `heroku.yml` file is correctly set up for Heroku's container runtime.

Heroku's container runtime is different from Docker in that it doesn't support `docker-compose.yml` files. Instead, it uses `heroku.yml` for defining the app's services.

In your case, the `heroku.yml` file is already set up to build the Dockerfile and run the web service with `npm start`.

However, the `release` command in the `heroku.yml` file is not correctly set up. The `release` command is used to specify a command to run before a new release of your app is deployed.

Here are the *SEARCH/REPLACE* blocks:

lethal-scrapbook/heroku.yml
```yaml
<<<<<<< SEARCH
release:
  image: web
release:
  image: web
=======
release:
  command:
    - ./init-db.sh
>>>>>>> REPLACE
```

This change will ensure that the `init-db.sh` script is run before a new release of your app is deployed, which will initialize your database.

Also, the `CMD` command in the Dockerfile should be changed to start the server instead of running the `init-db.sh` script, because the script will now be run by the `release` command in the `heroku.yml` file.

lethal-scrapbook/Dockerfile
```Dockerfile
<<<<<<< SEARCH
CMD ["./init-db.sh"]
=======
CMD ["npm", "start"]
>>>>>>> REPLACE
```

This change will ensure that the Docker container starts the server when it is run.